style: move aoss client specific stuff from vector store to example
